Add `CustomDump` trait
