+++ Package@swift-5.9.swift
@@ -56,6 +56,7 @@ let package = Package(
 for target in package.targets {
   target.swiftSettings = target.swiftSettings ?? []
   target.swiftSettings?.append(contentsOf: [
+    .define("CustomDump"),
     .define("IdentifiedCollections"),
   ])
 }
