+++ Tests/SharingTests/SharedTests.swift
@@ -1,6 +1,7 @@
 import CustomDump
 import Dependencies
 import Foundation
+import IdentifiedCollections
 import PerceptionCore
 import Sharing
 import Testing
